openharmony-schedule - 13

+++ openharmony/arkui_ace_engine/frameworks/core/components_ng/pattern/swiper/arc_swiper_pattern.cpp
@@ -243,13 +243,13 @@ void ArcSwiperPattern::PlayVerticalAnimation(const OffsetF& offset, int32_t inde
 bool ArcSwiperPattern::IsPreItem(int32_t index, float translate, bool rollback)
 {
     if (translate < 0) {
-        if (index < itemPosition_.size() / HALF) {
+        if (index < static_cast<int32_t>(itemPosition_.size() / HALF)) {
             return !rollback;
         } else {
             return rollback;
         }
     } else {
-        if (index < itemPosition_.size() / HALF) {
+        if (index < static_cast<int32_t>(itemPosition_.size() / HALF)) {
             return rollback;
         } else {
             return !rollback;
